Added `ExpressionBase`
