+++ src/NCalc.Async/AsyncExpression.cs
@@ -15,44 +15,8 @@ namespace NCalc;
 /// It supports caching, custom parameter and function evaluation, and options for handling null parameters and iterating over parameter collections.
 /// The class manages the parsing, validation, and evaluation of expressions, and provides mechanisms for error detection and reporting.
 /// </summary>
-public class AsyncExpression
+public class AsyncExpression : ExpressionBase<AsyncExpressionContext>
 {
-    /// <summary>
-    /// Static property to enable or disable cache.
-    /// Default Value: True
-    /// </summary>
-    public static bool CacheEnabled { get; set; } = true;
-    
-    /// <summary>
-    /// Options for the expression evaluation.
-    /// </summary>
-    public ExpressionOptions Options
-    {
-        get => Context.Options;
-        set => Context.Options = value;
-    }
-
-    /// <summary>
-    /// Culture information for the expression evaluation.
-    /// </summary>
-    public CultureInfo CultureInfo
-    {
-        get => Context.CultureInfo;
-        set => Context.CultureInfo = value;
-    }
-    
-    protected AsyncExpressionContext Context { get; init; }
-
-    /// <summary>
-    /// Parameters for the expression evaluation.
-    /// </summary>
-    public IDictionary<string, object?> Parameters
-    {
-        get => Context.StaticParameters;
-        set => Context.StaticParameters = value;
-    }
-
-
     public IDictionary<string, AsyncExpressionParameter> DynamicParameters
     {
         get => Context.DynamicParameters;
@@ -83,29 +47,14 @@ public class AsyncExpression
         add => Context.AsyncEvaluateParameterHandler += value;
         remove => Context.AsyncEvaluateParameterHandler -= value;
     }
-    
-    /// <summary>
-    /// Textual representation of the expression.
-    /// </summary>
-    public string? ExpressionString { get; protected init; }
 
-    public LogicalExpression? LogicalExpression { get; protected set; }
-
-    public Exception? Error { get; private set; }
-
-    protected ILogicalExpressionCache LogicalExpressionCache { get;}
-
-    protected ILogicalExpressionFactory LogicalExpressionFactory { get; }
 
     protected IAsyncEvaluationService EvaluationService { get; }
     
     
-    private AsyncExpression(AsyncExpressionContext? context = null)
+    private AsyncExpression(AsyncExpressionContext? context = null) : base(context ?? new AsyncExpressionContext())
     {
-        LogicalExpressionCache = Cache.LogicalExpressionCache.GetInstance();
-        LogicalExpressionFactory = Factories.LogicalExpressionFactory.GetInstance();
         EvaluationService = new AsyncEvaluationService();
-        Context = context ?? new();
     }
 
     public AsyncExpression(
@@ -113,13 +62,9 @@ public class AsyncExpression
         AsyncExpressionContext context,
         ILogicalExpressionFactory factory,
         ILogicalExpressionCache cache,
-        IAsyncEvaluationService evaluationService)
+        IAsyncEvaluationService evaluationService) : base(expression,context,factory,cache)
     {
-        ExpressionString = expression;
-        Context = context;
-        LogicalExpressionCache = cache;
         EvaluationService = evaluationService;
-        LogicalExpressionFactory = factory;
     }
     
     public AsyncExpression(
@@ -127,23 +72,18 @@ public class AsyncExpression
         AsyncExpressionContext context,
         ILogicalExpressionFactory factory,
         ILogicalExpressionCache cache,
-        IAsyncEvaluationService evaluationService)
+        IAsyncEvaluationService evaluationService) : base(logicalExpression, context,factory,cache)
     {
-        LogicalExpression = logicalExpression;
-        Context = context;
-        LogicalExpressionCache = cache;
         EvaluationService = evaluationService;
-        LogicalExpressionFactory = factory;
     }
-    
+
     public AsyncExpression(string expression, AsyncExpressionContext? context = null) : this(context)
     {
         ExpressionString = expression;
     }
-
-
+    
     // ReSharper disable once RedundantOverload.Global
-    // Reason: False positive, ExpressionContext have implicit conversions.
+    // Reason: False positive, AsyncExpressionContext have implicit conversions.
     public AsyncExpression(string expression) : this(expression, ExpressionOptions.None)
     {
     }
@@ -160,7 +100,7 @@ public class AsyncExpression
     }
 
     // ReSharper disable once RedundantOverload.Global
-    // Reason: False positive, ExpressionContext have implicit conversions.
+    // Reason: False positive, AsyncExpressionContext have implicit conversions.
     public AsyncExpression(LogicalExpression logicalExpression) : this(logicalExpression, ExpressionOptions.None)
     {
     }
@@ -170,55 +110,6 @@ public class AsyncExpression
     {
     }
     
-    private LogicalExpression? GetLogicalExpression()
-    {
-        if (string.IsNullOrEmpty(ExpressionString))
-            throw new NCalcException("Expression cannot be null or empty.");
-
-        var isCacheEnabled = CacheEnabled && !Options.HasFlag(ExpressionOptions.NoCache);
-
-        LogicalExpression? logicalExpression = null;
-
-        if (isCacheEnabled && LogicalExpressionCache.TryGetValue(ExpressionString!, out logicalExpression))
-            return logicalExpression!;
-
-        try
-        {
-            logicalExpression = LogicalExpressionFactory.Create(ExpressionString!, Context.Options);
-            if (isCacheEnabled)
-                LogicalExpressionCache.Set(ExpressionString!, logicalExpression);
-        }
-        catch (Exception exception)
-        {
-            Error = exception;
-        }
-
-        return logicalExpression;
-    }
-    
-    /// <summary>
-    /// Create the LogicalExpression in order to check syntax errors.
-    /// If errors are detected, the Error property contains the exception.
-    /// </summary>
-    /// <returns>True if the expression syntax is correct, otherwise False.</returns>
-    [MemberNotNullWhen(true, nameof(Error))]
-    public bool HasErrors()
-    {
-        try
-        {
-            LogicalExpression = LogicalExpressionFactory.Create(ExpressionString!, Context.Options);
-
-            // In case HasErrors() is called multiple times for the same expression
-            return LogicalExpression != null && Error != null;
-        }
-        catch (Exception exception)
-        {
-            Error = exception;
-            return true;
-        }
-    }
-    
-
     /// <summary>
     /// Asynchronously evaluates the logical expression.
     /// </summary>
@@ -259,31 +150,5 @@ public class AsyncExpression
         }
 
         return results;
-    }
-    
-    /// <summary>
-    /// Returns a list with all parameter names from the expression.
-    /// </summary>
-    public List<string> GetParameterNames()
-    {
-        var parameterExtractionVisitor = new ParameterExtractionVisitor();
-        LogicalExpression ??= LogicalExpressionFactory.Create(ExpressionString!, Context.Options);
-        return LogicalExpression.Accept(parameterExtractionVisitor);
-    }
-
-    [Obsolete("Please use GetParameterNames (correct english spelling).")]
-    public List<string> GetParametersNames()
-    {
-        return GetParameterNames();
-    }
-    
-    /// <summary>
-    /// Returns a list with all function names from the expression.
-    /// </summary>
-    public List<string> GetFunctionNames()
-    {
-        var functionExtractionVisitor = new FunctionExtractionVisitor();
-        LogicalExpression ??= LogicalExpressionFactory.Create(ExpressionString!, Context.Options);
-        return LogicalExpression.Accept(functionExtractionVisitor);
     }
 }
