+++ src/NCalc.Sync/Expression.cs
@@ -15,14 +15,8 @@ namespace NCalc;
 /// It supports caching, custom parameter and function evaluation, and options for handling null parameters and iterating over parameter collections.
 /// The class manages the parsing, validation, and evaluation of expressions, and provides mechanisms for error detection and reporting.
 /// </summary>
-public partial class Expression
+public partial class Expression : ExpressionBase<ExpressionContext>
 {
-    /// <summary>
-    /// Static property to enable or disable cache.
-    /// Default Value: True
-    /// </summary>
-    public static bool CacheEnabled { get; set; } = true;
-    
     /// <summary>
     /// Event triggered to handle function evaluation.
     /// </summary>
@@ -41,34 +35,6 @@ public partial class Expression
         remove => Context.EvaluateParameterHandler -= value;
     }
     
-    /// <summary>
-    /// Options for the expression evaluation.
-    /// </summary>
-    public ExpressionOptions Options
-    {
-        get => Context.Options;
-        set => Context.Options = value;
-    }
-
-    /// <summary>
-    /// Culture information for the expression evaluation.
-    /// </summary>
-    public CultureInfo CultureInfo
-    {
-        get => Context.CultureInfo;
-        set => Context.CultureInfo = value;
-    }
-
-    protected ExpressionContext Context { get; init; }
-
-    /// <summary>
-    /// Parameters for the expression evaluation.
-    /// </summary>
-    public IDictionary<string, object?> Parameters
-    {
-        get => Context.StaticParameters;
-        set => Context.StaticParameters = value;
-    }
     
     public IDictionary<string, ExpressionParameter> DynamicParameters
     {
@@ -81,26 +47,12 @@ public partial class Expression
         get => Context.Functions;
         set => Context.Functions = value;
     }
-
-    /// <summary>
-    /// Textual representation of the expression.
-    /// </summary>
-    public string? ExpressionString { get; protected init; }
-
-    public LogicalExpression? LogicalExpression { get; protected set; }
-
-    public Exception? Error { get; private set; }
-
-    protected ILogicalExpressionCache LogicalExpressionCache { get; }
-    protected ILogicalExpressionFactory LogicalExpressionFactory { get; }
+    
     protected IEvaluationService EvaluationService { get; }
     
-    private Expression(ExpressionContext? context = null)
+    private Expression(ExpressionContext? context = null) : base(context ?? new ExpressionContext())
     {
-        LogicalExpressionCache = Cache.LogicalExpressionCache.GetInstance();
-        LogicalExpressionFactory = Factories.LogicalExpressionFactory.GetInstance();
         EvaluationService = new EvaluationService();
-        Context = context ?? new();
     }
 
     public Expression(
@@ -108,13 +60,9 @@ public partial class Expression
         ExpressionContext context,
         ILogicalExpressionFactory factory,
         ILogicalExpressionCache cache,
-        IEvaluationService evaluationService)
+        IEvaluationService evaluationService) : base(expression,context,factory,cache)
     {
-        ExpressionString = expression;
-        Context = context;
-        LogicalExpressionCache = cache;
         EvaluationService = evaluationService;
-        LogicalExpressionFactory = factory;
     }
     
     public Expression(
@@ -122,13 +70,9 @@ public partial class Expression
         ExpressionContext context,
         ILogicalExpressionFactory factory,
         ILogicalExpressionCache cache,
-        IEvaluationService evaluationService)
+        IEvaluationService evaluationService) : base(logicalExpression, context,factory,cache)
     {
-        LogicalExpression = logicalExpression;
-        Context = context;
-        LogicalExpressionCache = cache;
         EvaluationService = evaluationService;
-        LogicalExpressionFactory = factory;
     }
 
     public Expression(string expression, ExpressionContext? context = null) : this(context)
@@ -163,28 +107,7 @@ public partial class Expression
         CultureInfo? cultureInfo = null) : this(logicalExpression, new ExpressionContext(options, cultureInfo))
     {
     }
-
-    /// <summary>
-    /// Create the LogicalExpression in order to check syntax errors.
-    /// If errors are detected, the Error property contains the exception.
-    /// </summary>
-    /// <returns>True if the expression syntax is correct, otherwise False.</returns>
-    [MemberNotNullWhen(true, nameof(Error))]
-    public bool HasErrors()
-    {
-        try
-        {
-            LogicalExpression = LogicalExpressionFactory.Create(ExpressionString!, Context.Options);
-
-            // In case HasErrors() is called multiple times for the same expression
-            return LogicalExpression != null && Error != null;
-        }
-        catch (Exception exception)
-        {
-            Error = exception;
-            return true;
-        }
-    }
+    
 
     /// <summary>
     /// Evaluates the logical expression.
@@ -208,32 +131,6 @@ public partial class Expression
         return EvaluationService.Evaluate(LogicalExpression!, Context);
     }
 
-    private LogicalExpression? GetLogicalExpression()
-    {
-        if (string.IsNullOrEmpty(ExpressionString))
-            throw new NCalcException("Expression cannot be null or empty.");
-
-        var isCacheEnabled = CacheEnabled && !Options.HasFlag(ExpressionOptions.NoCache);
-
-        LogicalExpression? logicalExpression = null;
-
-        if (isCacheEnabled && LogicalExpressionCache.TryGetValue(ExpressionString!, out logicalExpression))
-            return logicalExpression!;
-
-        try
-        {
-            logicalExpression = LogicalExpressionFactory.Create(ExpressionString!, Context.Options);
-            if (isCacheEnabled)
-                LogicalExpressionCache.Set(ExpressionString!, logicalExpression);
-        }
-        catch (Exception exception)
-        {
-            Error = exception;
-        }
-
-        return logicalExpression;
-    }
-
     private List<object?> IterateParameters()
     {
         var parameterEnumerators = ParametersHelper.GetEnumerators(Parameters, out var size);
@@ -252,31 +149,5 @@ public partial class Expression
         }
 
         return results;
-    }
-
-    /// <summary>
-    /// Returns a list with all parameter names from the expression.
-    /// </summary>
-    public List<string> GetParameterNames()
-    {
-        var parameterExtractionVisitor = new ParameterExtractionVisitor();
-        LogicalExpression ??= LogicalExpressionFactory.Create(ExpressionString!, Context.Options);
-        return LogicalExpression.Accept(parameterExtractionVisitor);
-    }
-    
-    [Obsolete("Please use GetParameterNames (correct english spelling).")]
-    public List<string> GetParametersNames()
-    {
-        return GetParameterNames();
-    }
-    
-    /// <summary>
-    /// Returns a list with all function names from the expression.
-    /// </summary>
-    public List<string> GetFunctionNames()
-    {
-        var functionExtractionVisitor = new FunctionExtractionVisitor();
-        LogicalExpression ??= LogicalExpressionFactory.Create(ExpressionString!, Context.Options);
-        return LogicalExpression.Accept(functionExtractionVisitor);
     }
 }
